requirements.txt
# Core AgentCore and Strands
bedrock-agentcore>=1.1.2
strands-agents>=1.20.0
bedrock-agentcore-starter-toolkit>=0.2.5

# Web framework
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
websockets>=15.0.1

# AWS SDK
boto3>=1.42.17
botocore>=1.42.17
# Note: aws-sdk-bedrock-runtime requires Python 3.12+, removed for Python 3.11 compatibility
# boto3 and bedrock-agentcore provide the necessary Bedrock functionality

# Tools and utilities
# strands-tools is included in strands-agents package (use: from strands.tools import tool)
pydantic>=2.12.5
python-dotenv>=1.2.1
requests>=2.32.5

# Audio processing
numpy>=2.4.0
soundfile>=0.13.1
pyaudio>=0.2.14

# Google OAuth2
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
PyJWT>=2.8.0
click>=8.1.0

# Testing
pytest>=9.0.2
pytest-asyncio>=1.3.0
pytest-cov>=7.0.0
httpx>=0.28.1

# Development
black>=25.12.0
ruff>=0.14.10
mypy>=1.19.1
